[쇼핑 주문 3, 4 단계] 디랙 미션 제출합니다. (#146)
* refactor: Coroutine 적용

* refactor: Coroutine 적용

* docs: 기능 목록 작성

* feat: 결제 화면 XML 추가

* feat: 쿠폰 XML 추가

* refactor: 패키지 구조 변경

* feat: 쿠폰 관련 Service, DataSource, Repository 추가

* feat: 상품 추천 화면에서 주문하기 클릭 시 결제 화면으로 이동하는 기능 구현

* refactor: 쿠폰 모델 수정

* feat: CheckoutViewModel 추가

* feat: 결제 화면에서 쿠폰 목록 표시

* refactor: 결제 화면에서 선택된 상품들의 아이디 목록을 받을 수 있도록 변경

* feat: 주문 화면으로 선택된 상품 정보 전달 및 주문 화면에서 주문 금액을 표시하는 기능 추가

* refactor: 선택 가능한 쿠폰을 한 개로 제한

* feat: 쿠폰이 사용 가능한지 판별하는 함수 구현

* feat: 쿠폰별 사용 가능 여부를 반환하는 함수 및 할인 금액을 반환하는 함수 구현

* refactor: 사용 가능한 쿠폰만 표시되도록 변경

* refactor: 결제 화면에서 배송비가 표시되도록 변경

* refactor: 추천 상품을 장바구니에 추가하는 로직 변경

* refactor: 선택된 상품이 없을 시 주문을 할 수 없도록 변경

* refactor: FreeShippingCoupon을 data class로 변경

* feat: 총 결제 금액을 표시하는 기능 추가

* refactor: 결제 화면 레이아웃 수정

* refactor: 최소 주문 금액이 없는 쿠폰은 최소 주문 금액이 표시되지 않도록 변경

* refactor: 쿠폰 할인 금액이 마이너스로 표시되도록 변경

* refactor: 쿠폰 만료일이 표시되도록 변경 및 최소 주문 금액 표기 변경

* feat: 결제 화면에 뒤로가기 추가

* feat: 결제하기 기능 구현

* refactor: DetailViewModel에서 productId를 생성자에서 받도록 변경

* refactor: DetailViewModel에서 productId를 생성자에서 받도록 변경

* refactor: CheckoutViewModel에서 선택된 상품의 ID 목록을 생성자에서 받도록 변경

* refactor: CheckoutViewModel에서 선택된 상품의 ID 목록을 생성자에서 받도록 변경

* fix: 주문 완료 시 주문한 장바구니 상품이 모두 제거되지 않는 현상 수정

* refactor: 가격 관련 변수들을 Long으로 변경

* fix: 장바구니 전체 선택 시 현재 페이지에 있는 상품만 선택되는 현상 수정

* fix: 사용 가능한 쿠폰이 정상적으로 로드되지 않는 현상 수정

* refactor: 주문 API 연결

* fix: 결제 시 구매한 상품이 장바구니에서 정상적으로 제거되지 않는 현상 수정

* refactor: 레이아웃 수정

* fix: 장바구니 화면에서 상품 전체 선택 시 한 페이지에 상품이 5개를 초과해 표시될 수 있는 현상 수정

* docs: 기능 목록 갱신

* fix: 추천 상품 화면에서 상품이 장바구니에 정상적으로 추가되지 않는 현상 수정

* refactor: 불필요한 로그 제거

* refactor: 불필요한 clickable 및 focusable 속성 제거

* refactor: 쿠폰 할인 함수 이름 변경

* fix: 장바구니에서 선택된 상품의 수량을 변경해도 총 액수 및 선택된 개수 표시에 반영되지 않는 현상 수정

* refactor: CartFragment에서 사용하지 않는 LiveData observe 제거

* refactor: 장바구니에 상품 추가 시 추가된 상품 정보를 반환하도록 변경

* refactor: DetailViewModel의 장바구니 담기 로직을 CartRepository로 이동

* refactor: XML 내의 스트링 포맷 방식 변경

* refactor: 결제 화면 레이아웃 변경

* refactor: 상품 로딩 로직 변경 및 함수명 변경

* refactor: 장바구니 체크박스 BindingAdapter에서 불필요한 속성 제거

* refactor: 상품 선택 이벤트 핸들러 파라미터 변경

* refactor: 쿠폰 선택 시 CouponUiModel의 isSelected 상태를 사용하도록 변경

* refactor: 토스트 메시지 리소스 처리

* refactor: DataSource에서 Domain Model을 반환하도록 변경

* refactor: 주문 완료 시 장바구니 상품을 제거하는 로직을 OrderRepository로 이동

* refactor: 패키지 구조 변경 및 클래스 이름 변경

* refactor: CartRepository에 ID의 목록으로 장바구니 상품의 목록을 반환하는 함수 추가

* refactor: 백 버튼 콜백에 viewLifecycleOwner 추가

* refactor: 상품 목록 화면에서 상품 목록을 불러오는 로직 변경

+++ app/src/main/java/woowacourse/shopping/presentation/model/ProductUiModel.kt
@@ -1,8 +1,8 @@
 package woowacourse.shopping.presentation.model
 
-import woowacourse.shopping.domain.model.CartItem
-import woowacourse.shopping.domain.model.Price
-import woowacourse.shopping.domain.model.Product
+import woowacourse.shopping.domain.CartItem
+import woowacourse.shopping.domain.Price
+import woowacourse.shopping.domain.Product
 
 data class ProductUiModel(
     val cartId: Long = 0L,
@@ -10,7 +10,7 @@ data class ProductUiModel(
     val category: String = "",
     val name: String,
     val imageUrl: String,
-    val price: Int,
+    val price: Long,
     val quantity: Int = 0,
 )
